Log graphql execution times
This will allow us to dig a bit deeper into some of the slow graphql
queries. Are there particular documents / base paths that are
reproducibly slow?

Ruby's Benchmark#measure method returns a [Benchmark::Tms](https://ruby-doc.org/3.4.1/stdlibs/benchmark/Benchmark/Tms.html)
which has various timings. The main ones we're interested in here are
utime (user time), stime (system time) and real (elapsed real time).
We're not so interested in cstime or cutime because we don't use child
processes in this block.

I considered trying to instrument this with tracing instead, but I think
the sample rate currently has to be set for the entire application, and
it's too low to be useful at the moment, and there's too much other
stuff happening in this app for me to want to risk increasing it.