Send the content block’s update type when queuing
Before, we were sending a “special” update type of `host_content`,
which makes it difficult to work out if a content block has been
updated with a major or minor update type. This removes the call,
and just sends `major` or `minor`.

In future commits, we’ll add an alternative method of working out
if the source edition is a content block.