Pass triggering edition to the DownstreamPayload
If a republish has been triggered by a linked edition, then we
pass this to the downstream payload. This will allow us to override the
"updated at" timestamp to that of the embedded content change rather
than that of the host edition change, which we’ll do in a future commit.