Always update dependencies for content blocks
`update_dependencies` is only set to `true` if the `edition_diff` class
notices any changes. Content Blocks often have a much more complex
makeup, so it’s harder to diff these items. Additionally, updating a
content block should always trigger a republish, as we’ll always want
to re-render the blocks and send the updates to the content store.