Make titles optional
The initial thinking behind making sure embedded objects had titles was
because we needed a title to generate a slug for embed codes. We’ve
since worked out that we can generate an automatic slug if the title is
missing, so let’s keep title around as a default attribute, but make
it optional.