Always return symbolized keys
We currently have a mixture of symbols and string for keys in the hash
returned by `Queries::GetLinkChanges`, since `slice` always return
string keys.

Making this consistent, so there are symbols used throughout, which
makes the tests look neater. It makes no difference downstream, as the
result is rendered as JSON.