Add indices to `link_changes` table
We are querying on various fields from this table, but there are no
indices. This means the query in `Queries::GetLinkChanges` to retrieve
link changes takes around ~8000ms. This reduces the same query to ~6ms.

This should resolve an issue where looking up change history in Content
Tagger times out as the request takes longer than the nginx timeout.