Add 'flexible_page' schema
"Anything goes" in the flexible page schema. Instead, the schema
and validation will live in Whitehall.

Part of the motivation for 'flexible pages' is to be able to
quickly respond to new user needs, e.g. building new page types
like for Coronavirus or Brexit. If Frontend is built flexibly
enough, then the only repo that would need any changes in such
an event would be Whitehall (and perhaps even then there may be
no changes required in Whitehall itself, depending on how self-serve
we make the offering). We want to avoid having to raise PRs
across Whitehall, Publishing API and Frontend for any tweaks to
flexible pages.

This commit is a proof of concept while we explore the idea
further.