Use the query in the presenter
This uses the `Presenters::Queries::ChangeHistory` query to fetch all
change notes for an edition, as well as any change notes for any
embedded content blocks.

If change notes already exist in the details hash, we only fetch the
change notes for linked editions, and merge these into the details
hash, ordering by timestamp and converting all timestamps to UTC, to
ensure consistency.