Add GraphqlQueryBuilder, fragments and script
GraphqlQueryBuilder and the fragments were initially copied verbatim
from the govuk-graphql repository.

govuk-graphql's version of the query builder took in a single content
item and derived a GraphQL query based on that item's fields.

In reality, the fields present on content items can vary quite a bit
even between items of the same type (schema_name). So we've adapted the
existing script to take in multiple content items of the same type and
merge them all together before building (what we hope is) a canonical
GraphQL query that will include all relevent fields.

The script assumes the existence of a local Content Store repository
that has downloaded content item files present.

We could also consider extending this to build additional fragments.