Add GraphqlQueryBuilder and script
GraphqlQueryBuilder was copied verbatim from the govuk-graphql
repository.

govuk-graphql's version of the query builder took in a single content
item and derived a GraphQL query based on that item's fields.

In reality, the fields present on content items can vary quite a bit
even between items of the same type (schema_name). So we've adapted the
existing script to take in the content type's schema (from
content_schemas/), some data from ExpansionRules and a random example
content item from GovukSchemas. Hopefully this is getting us close to
building a canonical GraphQL query including All The Fields that the
given content type can present.

In transit from govuk-graphql, we dropped the use of predefined GraphQL
fragments since we weren't 100% sure they were exhaustive. We plan to
extract our own later on from the queries that this script produces.