Add Content Store-like API endpoint for GraphQL
This will have the same API inputs and outputs as Content Store, but
will be powered by GraphQL. It replicates the behaviour (and tests) from
Content Store's `ContentItemsController`, albeit limited to only live
content.

Doing this will allow us to make a 'drop-in' replacement for Content
Store, as no changes will need to be made to frontend applications
(other the the URL for the API).

Notes:
- `max_cache_time` has been skipped as none of the currently supported
  formats include this field. Support will need to be added as the
  number of schemas is increased.
- There is no support for 'exact' and 'prefix' routes. These are not
  used by the content types we currently support in GraphQL.
- The default cache times have been copied from Content Store, to ensure
  public Content Store-like responses have the same cache headers as
  those currentl provided by Content Store.