Seed srand with RSpec's seed, use rand for schemas
RSpec's seed is the same number for an entire test run, so if we use it
for RandomExample.new we'll get the same example every single time,
which makes `generate_random_edition` a badly named method.

Instead, we seed Ruby's default random number generator (RNG) using
srand with RSpec's seed. We then use Ruby's RNG to generate
deterministic seeds for RandomExample, so it gets a different seed each
time it's called, but they come from a deterministic source.

This means generate_random_edition will return a different random
edition each time it's called, but if RSpec is called with the same seed
it will return the same sequence of random editions each time.