Render change history timestamps as ISO8601
We were calling `.to_s` on the activesupport datetime objects before, which returns dates like:

2025-08-15 17:27 UTC

(Note: spaces, ends with UTC)

The schemas expect ISO8601 formatted times, like:

2025-08-15T17:27Z

(Note: T instead of space, ends with Z)

... so this is actually broken.

Annoyingly, this was only caught by the random message queue test, which is flaky for other reasons (to be fixed in subsequent commits). This made it very difficult to track this down.