Seed random examples using RSpec's seed
One problem with using random examples is that it can be hard to
reproduce failures.

For example, we currently have a problem where if an example is
generated with document_type: redirect (1 in 191) it will put a
different kind of message on the queue, breaking the tests.

Replicating this is very frustrating, because you have to run the tests
almost 200 times before you can replicate the failure.

Like with randomly ordered tests, it would be nice if we could specify a
seed, and derive the random examples from that seed. That way, if the
tests fail with a random issue, it can be reproduced by re-running the
tests with the same seed.

We need to ensure that repeated calls to generate_random_edition produce
different random editions, because we call it in a loop 50.times and it
would be a waste of time if it returned the same edition every time. To
achieve this, we extract a seeded RandomExample instance, and then use
it repeatedly (each call to payload will produce a new random example,
but they'll be deterministic relative to the seed of the entire test
run).

We also print RSpec's seed in spec_helper, because it isn't printed by
default unless the tests are being run in a random order (which we don't
do in this app, yet).