Don't return generic errors from GraphQL
If an invalid GraphQL query is executed against the schema, we get an
object containing errors back. For example, if we request a field that's
not defined on one of our GraphQL types (like Edition or Details), we
get undefined field errors

We previously weren't checking for this anywhere, so we'd end up
returning the error object with a 200 response from the
GraphqlController. Hopefully we wouldn't roll out
`GraphqlController#live_content` support for schemas with invalid
queries, but this change catches any such cases by raising instead of
returning a non-content item object