Remove nested GraphQL types from "details" field
In an effort to make our lives easier, we're redefining all of the
non-primitive types inside "details" to be just plain JSON fields
instead of custom-typed objects (e.g. `Image`).

We're currently trying to automate building GraphQL queries from content
items and content schemas, and this change should make the code we're
writing a bit less complex since it won't have to care about selecting
fields beyond the first level of "details" fields.

Note: we definitely do still need to define that first level of fields
within "details" because, for example, there are lot of parts of
a content item's details that shouldn't be included when that content is
included as a link.

Also note: the changes to the spec assertions do reflect the changes
that'll be visible in the production service. We're loosening our
control over which nested fields will be included in the resulting
content item as well as the format of some of those fields (think
datetime strings).