WIP Remove reverse link types not in the database
It turns out that Content Schemas include a lot of irrelevant top-level
link types (that all happen to be reverse link types). We're unsure how
to definitively figure out which of these links to include in our
GraphQL query for a given content type, so we're experimenting with an
imperfect method: looking up existing links in the database that point
to our content type. If we don't find any, we remove the corresponding
link type from the GraphQL query.