Skip impossible reverse link types in query
`SchemaGenerator::ExpandedLinks` includes all reverse link types in all
frontend content schemas. Because our GraphQL query builder depends on
content schemas as its main source of information, those links are
included in every generated GraphQL query.

This only affects the top-level "links" of the content item being
requested, and not those links' links.

We've noticed that it's unlikely or impossible for the content type that
we're currently migrating to GraphQL (transaction) to have certain
reverse link types and therefore the query that our script is generating
is more bloated than it needs to be. We believe that this is applicable
to content types generally, not just the transaction type.

The reverse link types that we were confident adding rules for:

1. child_taxons -- this link type is used for relationships from one
   taxon to another
2. level_one_taxons -- this reverse link type belongs exclusively to the
   homepage content type, as defined in Content Tagger
3. ministers -- apparently the direct link type that this corresponds to
   has been retired https://github.com/alphagov/whitehall/commit/783365d36c8a592151f04831fd6300ea47d90004
4. policies -- this link is between a working_group and a policy, but
   policies have been retired, so we don't expect any changes to what's
   currently in the live data