Select details_json if we don't know which fields
When automatically generating a GraphQL query for a given content type,
the script sometimes encounters content items (e.g. links) with
"details" fields but without any named children. The source of this
problem is ExpansionRules. Where ExpansionRules says that a given
document_type and/or link_type should have "details" but doesn't name
any of details's children, the result appears to be that all of the
details will be presented to Content Store, so in our case, with
read-time link expansion, the equivalent is to select the new
details_json field.