Support diffing using local frontend apps
We want to be able to run frontend diffs locally so that we can test
freshly-generated GraphQL queries without deploying anything

The change in the bulk diff while loop was prompted by an issue whereby
only the first base path would be diffed. It's likely that one of the
changes elsewhere was swallowing standard in, so once it got to looping
there was nothing left. This change prevents that. See:
https://stackoverflow.com/a/13800476/4002016