Prevent eating up command line options
Previously if an command line option that takes a value was provided
without a corresponding value argument, it would eat up the next option
argument. This wasn't causing problems because we we only ever provided
the possibly value-less options at the end, and if one didn't have a
value the other wouldn't. But now we need to introduce another possibly
value-less option to support running the diffs against a dev
environment, so we should check whether the next argument is an option
before shifting it

We already call these options without values in the `prepare_html`
function - we only pass in username and password values in integration,
but always include the flags to keep the function simple