Exit diff on curl error
When running the diffs against a dev environment, if a particular
required frontend app isn't up the curl command would fail to resolve
the host. This would be the same when testing Publishing API and Content
Store so we'd get no diff, but the script would just continue onto the
next base path and the curl error could easily get missed without keen
observation of the script output. This is a failed diff, so we should
exit the script immediately, leaving the curl error visible, prompting
the running dev to fix the setup issue before trying again