Expose AvailableTranslations as Editions
This method is being added for the benefit of the GraphQL EditionType.

EditionType currently uses AvailableTranslations#translations to fetch
its "available_translations" links, but that method was designed for
presenting to Content Store and there are differences between the data
that we present to Content Store and the data that we want to make
available through the GraphQL API. For convenience, it's more
straightforward to give EditionType the Edition instances and let the
GraphQL schema and query decide which fields to use.

---

There could be an argument for extracting the database query from this
presenter, and sharing just the query between this presenter and the
EditionType. I'm not familiar enough yet with the design of this
codebase to know the right way to approach a change like that, so I'm
proposing just living with this awkwardness for a little while until
it's clearer how to organise it.