Don't compact GraphQL content item
This was the original motivation for creating this service, but we've
since found that it results in a content item that's not valid against
the schema for at least some content types

For instance, the world index requires a description. Its actual
description is nil, but to be valid against the schema we still need to
include the property. This service was removing the property altogether
because of the top-level compacting

In practice the frontend apps are unlikely to be relying on the
existence of keys whose values are nil, but we think that producing a
schema-valid response is preferable to a minimal one, barring any
performance issues

This issue was surfaced when trying to expand the compacting to links
and all nested objects. This led to more validation errors before closer
inspection revealed the issue with the existing processing

NOTE: this is a red commit. The live_content_spec won't pass with these
changes. Another change to this file is necessary to produce a valid
content item for news articles based on their current query - it feels
preferable to break that into a separate commit and update the
tests after

```rb
def edition
 deep_transform_hash(query_result.dig("data", "edition"))
end

def deep_transform_hash(hash)
 hash.map { deep_transform(_1.to_s, _2) }.compact.to_h
end

def deep_transform(key, value)
 case [key, value]
 in [String, Hash]
   new_value = deep_transform_hash(value)
   [key, new_value] unless new_value == {}
 in [String, [Hash, *]]
   new_value = value.map(&method(:deep_transform_hash))
   [key, new_value] unless new_value == []
 in [String, nil | "" | [] | {}]
   nil
 else
   [key, value]
 end
end
```