Allow travel_advice to have topical_events links
This link type was previously allowed in this content schema. It appears
to have been added originally as a temporary solution to a request
relating to "/foreign-travel-advice/afghanistan". It was eventually
removed (2 years ago, ce808314b34fc51ddee566823790693626cbf8a7), but:

1. the link wasn't removed from the content item, and
2. it seems that by that point other travel_advice already had
   topical_events links of their own, which also weren't removed

Adding this link type back into the content schema hopefully means that
those live content items with existing topical_events links will now be
considered valid against the schema. I'm assuming that those links have
been published intentionally, so I believe that this is the correct
outcome.

We're currently working on automating the building of GraphQL queries
for each content type. One of the sources of information that our script
is relying on is the content schema for the given content type. Without
topical_events links being included in the travel_advice content schema,
the script doesn't include them in the generated travel_advice GraphQL
query and that causes a difference between the GraphQL-backed content
item and Content Store's version which this change should also resolve.