Generate travel_advice.graphql query
This GraphQL query has been generated by running:

  `script/live_content/build_query travel_advice`

The query is quite large. It seems to be selecting a lot more stuff than
any of the existing, published travel_advice content actually has (e.g.
unlikely link types).

We could manually remove some of those fields from the query or we could
re-generate the query as the script is improved. But it won't
necessarily be possible to have both manual and automated improvements.

For now, the content items that this query produces (via the
/graphql/content endpoint) don't seem _too far_ from their equivalents
in Content Store (/api/content), and the pages that Frontend is
rendering backed by /graphql/content are identical to those backed by
Content Store's /api/content.

Changes in this commit other than the query file:

* change spec's example of a content type not supported by GraphQL to
  something else that still isn't supported
* add boilerplate to schema-validating spec to ensure FactoryBot creates
  an Edition with travel_advice's required properties present and in the
  correct format