Move query_result from constructor to method param
The GraphqlContentItemService gets initialized in the controller and
then we immediately call process. There's no real benefit to caching /
memoizing method calls, as we only call process one time.

It would be more convenient for testing if we could initialize a
GraphqlContentItemService once, and then call process on it with
different parameters to check different behaviour. It's probably not
possible to see the difference in this commit, because we don't take any
dependencies yet. I'd like to introduce some in future commits though.