Deep sort and stringify GraphQL content items
This is done in Content Store:
https://github.com/alphagov/content-store/pull/1100

Without this, we can get diffs like "Applies to England and Wales" in
Content Store and "Applies to Wales and England" due to the properties
of details["national_applicability"] being in a different order to how
Content Store returns them. This likely only affects fields with a JSON
type where we can't specify the order of fields (via selections)

We need to stringify the keys before sorting - otherwise symbols and
strings would be sorted separately. We wanted to stringify the keys
anyway for consistency...