Fetch required fields from the content schemas
I've used `fetch()` instead of dig for finding the required fields from
the schemas, as I figure it's better to be strict in what we expect and
handle an error from any schemas that don't conform to this structure
than to accidentally treat things as optional. The "details" object
doesn't always have required fields though (e.g. for the role schema),
so that one defaults to [].