Add an `order` field
This allows us to specify the order of embedded blocks within a
composed block with an array of strings in the format
`$block_type.$block_name`.

To (partly) ensure the orders are specified correctly, we specify
the embedded objects as a variable. We can then use the keys of
that object to ensure that the first part of the string contains
a block type. Unfortunately there doesn’t seem to be a way to validate
against the keys within those blocks, so we just add a regex to check
they are in the same format we expect object keys to be.