Deterministically order all linked editions
Even though we order links by link type and position everywhere, we
still have instances of links being ordered differently in our GraphQL
dataloaders than in the queries used by link expansion and dependency
resolution. This can cause both JSON and HTML diffs when comparing
GraphQL and Content Store

The reason this is possible is that links often have a position of 0
(about 80% of links, in fact) and in most queries we aren't specifying
how to order links with matching types and position. In such cases the
way Postgres plans the queries impacts the order in which we return
links, likely down to a difference in choice of index. The queries are
very distinct, so this is unsurprising

We think it's a good idea to have a deterministic ordering strategy in
general, and doing so will allow us to reduce what we consider to be
noise in the diffs (if two links have a position of 0, it shouldn't
matter which way they're ordered, but consistency is preferable)