Deterministically order all linked editions
Even though we order links by link type and position everywhere, we
still have instances of links being ordered differently in our GraphQL
dataloaders than in the queries used by link expansion and dependency
resolution. We're currently comparing JSON responses from GraphQL and
Content Store and how these are rendered as HTML in the frontend, and
the inconsistent ordering is causing noise in the diffs

We think this scenario only presents itself when retrieving reverse
links: the position should be distinct from other links of the same type
from the same source, but reverse links by their nature come from
multiple sources, so the positions in such cases are essentially random.
In such cases the way Postgres plans the queries impacts the order in
which we return links, likely down to a difference in choice of index.
The queries are very distinct, so this is unsurprising

We think it's a good idea to have a deterministic ordering strategy in
general, and doing so will allow us to reduce what we consider to be
noise in the diffs

We can't think of a scenario in which direct links of the same type from
the same source would have the same position, but just to be extra safe
we've included the ordering by link ID for direct link dataloader
queries too. In the link expansion code, the query is shared between
direct and reverse links

Co-authored-by: Mike Patrick <mike.patrick2@digital.cabinet-office.gov.uk>