Deduplicate linked-to editions in dataloader
In some cases, there's overlap between link set and edition links: there
might be one of each from the same source to the same target content ID.
Previously the fields we selected would have identical values for the
link set and edition link on each side of the UNION, and the UNION would
automatically filter out any such duplicate rows. However, we recently
started selecting the link ID (for sorting purposes) and since the
edition and link set link would have different link IDs, the rows would
no longer be exact copies. As a result, we'd retain both copies of the
edition

In the reverse link dataloader, we already included a check to prevent
inclusion of duplicates in the return value, so we're bringing the same
logic over here to deal with this new duplication issue

Co-authored-by: Ynda Jas <yndajas@gmail.com>