Deep sort (and stringify) GraphQL content items
This is done in Content Store:
https://github.com/alphagov/content-store/pull/1100

Without this, we can get diffs like "Applies to England and Wales" in
Content Store and "Applies to Wales and England" due to the properties
of details["national_applicability"] being in a different order to how
Content Store returns them. This likely only affects fields with a JSON
type where we can't specify the order of fields (via selections)

We stringify keys in the process. This stringifying allows us just to
use a vanilla `sort` on the result (no need to convert strings to
symbols or vice versa), but it also means we consistently use string
keys throughout the object returned by the service. Ultimately the
output of the `process` method will be converted to JSON where all keys
are strings, but having predictable keys will be useful in case we need
to do any more pre-JSON processing in future. Perhaps we should
stringify the keys as a separate step, but for now I've gone with
including it in the sorting process