BROKEN Remove documents.content_id from window function
This breaks the first two reverse link tests. I think we'll need to add
something to the window function in order to favour all the edition
links over all the link set links (e.g. if there are 2 edition links and
3 link set links, take the 2 edition links and dump all the link set
links). Sounds like Richard might know how to do this

---

With this, we were retaining some English editions where the equivalent
link expansion code (using Queries::EditionLinks, and overriding
Queries::Links) would not

Queries::EditionLinks#where probably has some clues here: the mode- and
locale-based conditions specifically. In contrast with Queries::Links,
which retains the English editions, Queries::EditionLinks takes
consideration of locales

Do we need to do similar for the direct links dataloader? In that case,
it looks like we should be concerned about documents.content_id and not
the target???