Do not log a lack of live edition to Sentry
At the moment, we are logging cases where there is no live edition (e.g.
the document only has draft editions) to Sentry. This is pointless as we
cannot take any action from these logs.

Looking at the tests, it appears the intention was to skip these
editions but it is unclear why they were being logged.

Since we do not use these logs for anything, this removes the logging to
Sentry and makes the tests more robust by checking that the content is
not pushed downstream (rather than using the reporting to Sentry as a
proxy with this check).