Omit suggested_ordered_related_items from GraphQL
Suggested ordered related item links have gradually become less useful
and are on the way out. They'll be removed from content schemas in a
separate piece of work with the data removed via a bulk republish

Given this context and suggested ordered related item links having been
a source of Content Store/GraphQL differences and complexity, we're
removing them from our GraphQL implementation. Here we remove them from
existing queries and update the query builder to omit them in future

I spent some time diffing a sample of editions of the four schema name
queries that are affected here and found no new diff as a result of this
change. It's possible that the rendering of these links has already been
removed from the frontend. As a result, there's nothing new to exclude
in the frontend diffs