Sample editions for GraphQL validation
This retrieves a sample of base paths, content IDs, and locales for use
in GraphQL validation scripts. It only samples live not-unpublished
editions. We have also sampled withdrawn unpublished editions in the
past, which we could add here later if useful, but they tend to vary
less by schema name than not-unpublished editions

We use `.pluck(...).sample(1000)` rather than `.limit(1000).pluck(...)`
so that the sample is random instead of the first N records. Based on
previous attempts, it seems that asking the database to order lots of
records randomly is more expensive than plucking a few fields from all
the records then using Ruby for the random sampling