Exit bulk frontend diff on HTTP error
Contrary to what the commit message for 493a875 might have suggested,
the `|| exit 1` wouldn't cause the script to exit if a prerequisite
frontend app wasn't running

The `|| exit 1` would cause the script to exit when given a base path
that couldn't be found in Publishing API's database. In such cases, the
frontend content path would have a domain of e.g. `null.dev.gov.uk` in
dev, which curl would be unable to resolve and then exit. I suppose this
could be indicative of Publishing API not being up, but it indicates
nothing about the status of the frontend app(s)

This adds the `--fail` flag to the `curl` command, which will cause curl
to return a non-zero exit code when the HTTP status is >=400, including
503 Service Unavailable. We now exit with the same code as curl if
there's an error. This avoids silent failure and a potentially false
negative reading of bulk frontend diff output